refactor: remove parse.go
